ai/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
dnspython==2.4.2
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
multidict==6.0.4
mypy==1.6.1
mypy-extensions==1.0.0
numpy==1.26.0
openai==0.28.1
pandas==2.1.1
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.6
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
tiktoken==0.5.1
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
yarl==1.9.2
pinecone-client==2.2.4
